refactor: DTO도 기능별로 패키지 분리 (findproject, teamup) - 코드 컨벤션에 따라 추후에 DTO 패키지명을 dto로 전면 수정할 필요가 있어보임 - exception도 마찬가지로 패키지 통일이 필요해보임
