+++ src/main/java/ssu/cromi/teamit/DTO/findproject/ProjectApplicationRequestDto.java
@@ -1,4 +1,4 @@
-package ssu.cromi.teamit.DTO;
+package ssu.cromi.teamit.DTO.findproject;
 
 import jakarta.validation.constraints.*;
 import java.util.List;
