+++ src/main/java/ssu/cromi/teamit/DTO/findproject/ProjectDetailResponseDto.java
@@ -1,4 +1,4 @@
-package ssu.cromi.teamit.DTO;
+package ssu.cromi.teamit.DTO.findproject;
 
 import lombok.Builder;
 import lombok.Getter;
