+++ src/main/java/ssu/cromi/teamit/controller/teamup/TeamController.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import ssu.cromi.teamit.DTO.CreateTeamRequestDto;
+import ssu.cromi.teamit.DTO.teamup.CreateTeamRequestDto;
 import ssu.cromi.teamit.security.UserDetailsImpl;
 import ssu.cromi.teamit.service.teamup.TeamService;
 import ssu.cromi.teamit.response.ApiResponse;
