+++ src/main/java/ssu/cromi/teamit/service/findproject/ProjectApplicationService.java
@@ -1,6 +1,6 @@
 package ssu.cromi.teamit.service.findproject;
 
-import ssu.cromi.teamit.DTO.ProjectApplicationRequestDto;
+import ssu.cromi.teamit.DTO.findproject.ProjectApplicationRequestDto;
 
 public interface ProjectApplicationService {
