+++ src/main/java/ssu/cromi/teamit/service/findproject/ProjectDetailService.java
@@ -1,6 +1,6 @@
 package ssu.cromi.teamit.service.findproject;
 
-import ssu.cromi.teamit.DTO.ProjectDetailResponseDto;
+import ssu.cromi.teamit.DTO.findproject.ProjectDetailResponseDto;
 
 public interface ProjectDetailService {
