+++ src/main/java/ssu/cromi/teamit/service/teamup/TeamService.java
@@ -1,6 +1,6 @@
 package ssu.cromi.teamit.service.teamup;
 
-import ssu.cromi.teamit.DTO.CreateTeamRequestDto;
+import ssu.cromi.teamit.DTO.teamup.CreateTeamRequestDto;
 
 public interface TeamService {
     Long createTeam(CreateTeamRequestDto requestDto, String userId);
