+++ src/main/java/ssu/cromi/teamit/service/teamup/TeamServiceImpl.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static ssu.cromi.teamit.util.EnumValidator.parseEnum;
-import ssu.cromi.teamit.DTO.CreateTeamRequestDto;
+import ssu.cromi.teamit.DTO.teamup.CreateTeamRequestDto;
 import ssu.cromi.teamit.entity.teamup.Project;
 import ssu.cromi.teamit.entity.teamup.ProjectMember;
 import ssu.cromi.teamit.entity.enums.*;
